server: split tests
