+++ pkg/server/tests/commontest/BUILD.bazel
@@ -1,0 +1,50 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "commontest_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "tidb_test.go",
+    ],
+    flaky = True,
+    shard_count = 47,
+    deps = [
+        "//pkg/config",
+        "//pkg/ddl/util",
+        "//pkg/extension",
+        "//pkg/metrics",
+        "//pkg/parser",
+        "//pkg/parser/ast",
+        "//pkg/parser/auth",
+        "//pkg/parser/mysql",
+        "//pkg/server",
+        "//pkg/server/internal/column",
+        "//pkg/server/internal/resultset",
+        "//pkg/server/internal/testserverclient",
+        "//pkg/server/internal/testutil",
+        "//pkg/server/internal/util",
+        "//pkg/server/tests/servertestkit",
+        "//pkg/session",
+        "//pkg/sessionctx/variable",
+        "//pkg/store/mockstore/unistore",
+        "//pkg/testkit",
+        "//pkg/testkit/testsetup",
+        "//pkg/util",
+        "//pkg/util/plancodec",
+        "//pkg/util/resourcegrouptag",
+        "//pkg/util/topsql",
+        "//pkg/util/topsql/collector",
+        "//pkg/util/topsql/collector/mock",
+        "//pkg/util/topsql/state",
+        "//pkg/util/topsql/stmtstats",
+        "@com_github_go_sql_driver_mysql//:mysql",
+        "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@com_github_tikv_client_go_v2//tikvrpc",
+        "@io_opencensus_go//stats/view",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
