+++ pkg/server/tests/commontest/tidb_test.go
@@ -12,25 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tests
+package commontest
 
 import (
 	"context"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/tls"
-	"crypto/x509"
-	"crypto/x509/pkix"
 	"database/sql"
 	"encoding/binary"
-	"encoding/pem"
 	"fmt"
 	"io"
-	"math/big"
 	"net"
-	"net/http"
 	"os"
-	"path/filepath"
 	"runtime"
 	"strings"
 	"sync"
@@ -43,9 +34,7 @@ import (
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/pkg/config"
 	ddlutil "github.com/pingcap/tidb/pkg/ddl/util"
-	"github.com/pingcap/tidb/pkg/domain"
 	"github.com/pingcap/tidb/pkg/extension"
-	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/parser"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/auth"
@@ -56,13 +45,12 @@ import (
 	"github.com/pingcap/tidb/pkg/server/internal/testserverclient"
 	"github.com/pingcap/tidb/pkg/server/internal/testutil"
 	util2 "github.com/pingcap/tidb/pkg/server/internal/util"
+	"github.com/pingcap/tidb/pkg/server/tests/servertestkit"
 	"github.com/pingcap/tidb/pkg/session"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
-	"github.com/pingcap/tidb/pkg/store/mockstore"
 	"github.com/pingcap/tidb/pkg/store/mockstore/unistore"
 	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/pingcap/tidb/pkg/util"
-	"github.com/pingcap/tidb/pkg/util/cpuprofile"
 	"github.com/pingcap/tidb/pkg/util/plancodec"
 	"github.com/pingcap/tidb/pkg/util/resourcegrouptag"
 	"github.com/pingcap/tidb/pkg/util/topsql"
@@ -75,133 +63,43 @@ import (
 	"go.opencensus.io/stats/view"
 )
 
-type tidbTestSuite struct {
-	*testserverclient.TestServerClient
-	tidbdrv *server2.TiDBDriver
-	server  *server2.Server
-	domain  *domain.Domain
-	store   kv.Storage
-}
-
-func createTidbTestSuite(t *testing.T) *tidbTestSuite {
-	cfg := util2.NewTestConfig()
-	cfg.Port = 0
-	cfg.Status.ReportStatus = true
-	cfg.Status.StatusPort = 0
-	cfg.Status.RecordDBLabel = true
-	cfg.Performance.TCPKeepAlive = true
-	return createTidbTestSuiteWithCfg(t, cfg)
-}
-
-func createTidbTestSuiteWithCfg(t *testing.T, cfg *config.Config) *tidbTestSuite {
-	ts := &tidbTestSuite{TestServerClient: testserverclient.NewTestServerClient()}
-
-	// setup tidbTestSuite
-	var err error
-	ts.store, err = mockstore.NewMockStore()
-	session.DisableStats4Test()
-	require.NoError(t, err)
-	ts.domain, err = session.BootstrapSession(ts.store)
-	require.NoError(t, err)
-	ts.tidbdrv = server2.NewTiDBDriver(ts.store)
-
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
-	require.NoError(t, err)
-	ts.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
-	ts.StatusPort = testutil.GetPortFromTCPAddr(server.StatusListenerAddr())
-	ts.server = server
-	ts.server.SetDomain(ts.domain)
-	ts.domain.InfoSyncer().SetSessionManager(ts.server)
-	go func() {
-		err := ts.server.Run()
-		require.NoError(t, err)
-	}()
-	ts.WaitUntilServerOnline()
-
-	t.Cleanup(func() {
-		if ts.domain != nil {
-			ts.domain.Close()
-		}
-		if ts.server != nil {
-			ts.server.Close()
-		}
-		if ts.store != nil {
-			require.NoError(t, ts.store.Close())
-		}
-		view.Stop()
-	})
-	return ts
-}
-
-type tidbTestTopSQLSuite struct {
-	*tidbTestSuite
-}
-
-func createTidbTestTopSQLSuite(t *testing.T) *tidbTestTopSQLSuite {
-	base := createTidbTestSuite(t)
-
-	ts := &tidbTestTopSQLSuite{base}
-
-	// Initialize global variable for top-sql test.
-	db, err := sql.Open("mysql", ts.GetDSN())
-	require.NoError(t, err)
-	defer func() {
-		err := db.Close()
-		require.NoError(t, err)
-	}()
-
-	dbt := testkit.NewDBTestKit(t, db)
-	topsqlstate.GlobalState.PrecisionSeconds.Store(1)
-	topsqlstate.GlobalState.ReportIntervalSeconds.Store(2)
-	dbt.MustExec("set @@global.tidb_top_sql_max_time_series_count=5;")
-
-	require.NoError(t, cpuprofile.StartCPUProfiler())
-	t.Cleanup(func() {
-		cpuprofile.StopCPUProfiler()
-		topsqlstate.GlobalState.PrecisionSeconds.Store(topsqlstate.DefTiDBTopSQLPrecisionSeconds)
-		topsqlstate.GlobalState.ReportIntervalSeconds.Store(topsqlstate.DefTiDBTopSQLReportIntervalSeconds)
-		view.Stop()
-	})
-	return ts
-}
-
 func TestRegression(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	if testserverclient.Regression {
 		ts.RunTestRegression(t, nil, "Regression")
 	}
 }
 
 func TestUint64(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	ts.RunTestPrepareResultFieldType(t)
 }
 
 func TestSpecialType(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	ts.RunTestSpecialType(t)
 }
 
 func TestPreparedString(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestPreparedString(t)
 }
 
 func TestPreparedTimestamp(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestPreparedTimestamp(t)
 }
 
 func TestErrorCode(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestErrorCode(t)
 }
 
 func TestAuth(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestAuth(t)
 	ts.RunTestIssue3682(t)
@@ -209,7 +107,7 @@ func TestAuth(t *testing.T) {
 }
 
 func TestIssues(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestIssue3662(t)
 	ts.RunTestIssue3680(t)
@@ -217,24 +115,24 @@ func TestIssues(t *testing.T) {
 }
 
 func TestDBNameEscape(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	ts.RunTestDBNameEscape(t)
 }
 
 func TestResultFieldTableIsNull(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestResultFieldTableIsNull(t)
 }
 
 func TestStatusAPI(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestStatusAPI(t)
 }
 
 func TestStatusPort(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	cfg := util2.NewTestConfig()
 	cfg.Port = 0
@@ -242,137 +140,13 @@ func TestStatusPort(t *testing.T) {
 	cfg.Status.StatusPort = ts.StatusPort
 	cfg.Performance.TCPKeepAlive = true
 
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.Error(t, err)
 	require.Nil(t, server)
 }
 
-func TestStatusAPIWithTLS(t *testing.T) {
-	ts := createTidbTestSuite(t)
-
-	dir := t.TempDir()
-
-	fileName := func(file string) string {
-		return filepath.Join(dir, file)
-	}
-
-	caCert, caKey, err := generateCert(0, "TiDB CA 2", nil, nil, fileName("ca-key-2.pem"), fileName("ca-cert-2.pem"))
-	require.NoError(t, err)
-	_, _, err = generateCert(1, "tidb-server-2", caCert, caKey, fileName("server-key-2.pem"), fileName("server-cert-2.pem"))
-	require.NoError(t, err)
-
-	cli := testserverclient.NewTestServerClient()
-	cli.StatusScheme = "https"
-	cfg := util2.NewTestConfig()
-	cfg.Port = cli.Port
-	cfg.Status.StatusPort = cli.StatusPort
-	cfg.Security.ClusterSSLCA = fileName("ca-cert-2.pem")
-	cfg.Security.ClusterSSLCert = fileName("server-cert-2.pem")
-	cfg.Security.ClusterSSLKey = fileName("server-key-2.pem")
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
-	require.NoError(t, err)
-	cli.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
-	cli.StatusPort = testutil.GetPortFromTCPAddr(server.StatusListenerAddr())
-	go func() {
-		err := server.Run()
-		require.NoError(t, err)
-	}()
-	time.Sleep(time.Millisecond * 100)
-
-	// https connection should work.
-	ts.RunTestStatusAPI(t)
-
-	// but plain http connection should fail.
-	cli.StatusScheme = "http"
-	//nolint:bodyclose
-	_, err = cli.FetchStatus("/status")
-	require.Error(t, err)
-
-	server.Close()
-}
-
-func TestStatusAPIWithTLSCNCheck(t *testing.T) {
-	ts := createTidbTestSuite(t)
-
-	dir := t.TempDir()
-
-	caPath := filepath.Join(dir, "ca-cert-cn.pem")
-	serverKeyPath := filepath.Join(dir, "server-key-cn.pem")
-	serverCertPath := filepath.Join(dir, "server-cert-cn.pem")
-	client1KeyPath := filepath.Join(dir, "client-key-cn-check-a.pem")
-	client1CertPath := filepath.Join(dir, "client-cert-cn-check-a.pem")
-	client2KeyPath := filepath.Join(dir, "client-key-cn-check-b.pem")
-	client2CertPath := filepath.Join(dir, "client-cert-cn-check-b.pem")
-
-	caCert, caKey, err := generateCert(0, "TiDB CA CN CHECK", nil, nil, filepath.Join(dir, "ca-key-cn.pem"), caPath)
-	require.NoError(t, err)
-	_, _, err = generateCert(1, "tidb-server-cn-check", caCert, caKey, serverKeyPath, serverCertPath)
-	require.NoError(t, err)
-	_, _, err = generateCert(2, "tidb-client-cn-check-a", caCert, caKey, client1KeyPath, client1CertPath, func(c *x509.Certificate) {
-		c.Subject.CommonName = "tidb-client-1"
-	})
-	require.NoError(t, err)
-	_, _, err = generateCert(3, "tidb-client-cn-check-b", caCert, caKey, client2KeyPath, client2CertPath, func(c *x509.Certificate) {
-		c.Subject.CommonName = "tidb-client-2"
-	})
-	require.NoError(t, err)
-
-	cli := testserverclient.NewTestServerClient()
-	cli.StatusScheme = "https"
-	cfg := util2.NewTestConfig()
-	cfg.Port = cli.Port
-	cfg.Status.StatusPort = cli.StatusPort
-	cfg.Security.ClusterSSLCA = caPath
-	cfg.Security.ClusterSSLCert = serverCertPath
-	cfg.Security.ClusterSSLKey = serverKeyPath
-	cfg.Security.ClusterVerifyCN = []string{"tidb-client-2"}
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
-	require.NoError(t, err)
-
-	cli.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
-	cli.StatusPort = testutil.GetPortFromTCPAddr(server.StatusListenerAddr())
-	go func() {
-		err := server.Run()
-		require.NoError(t, err)
-	}()
-	defer server.Close()
-	time.Sleep(time.Millisecond * 100)
-
-	hc := newTLSHttpClient(t, caPath,
-		client1CertPath,
-		client1KeyPath,
-	)
-	//nolint:bodyclose
-	_, err = hc.Get(cli.StatusURL("/status"))
-	require.Error(t, err)
-
-	hc = newTLSHttpClient(t, caPath,
-		client2CertPath,
-		client2KeyPath,
-	)
-	resp, err := hc.Get(cli.StatusURL("/status"))
-	require.NoError(t, err)
-	require.Nil(t, resp.Body.Close())
-}
-
-func newTLSHttpClient(t *testing.T, caFile, certFile, keyFile string) *http.Client {
-	cert, err := tls.LoadX509KeyPair(certFile, keyFile)
-	require.NoError(t, err)
-	caCert, err := os.ReadFile(caFile)
-	require.NoError(t, err)
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
-	tlsConfig := &tls.Config{
-		Certificates:       []tls.Certificate{cert},
-		RootCAs:            caCertPool,
-		InsecureSkipVerify: true,
-	}
-	tlsConfig.BuildNameToCertificate()
-	return &http.Client{Transport: &http.Transport{TLSClientConfig: tlsConfig}}
-}
-
 func TestMultiStatements(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunFailedTestMultiStatements(t)
 	ts.RunTestMultiStatements(t)
@@ -382,7 +156,7 @@ func TestSocketForwarding(t *testing.T) {
 	tempDir := t.TempDir()
 	socketFile := tempDir + "/tidbtest.sock" // Unix Socket does not work on Windows, so '/' should be OK
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	cli := testserverclient.NewTestServerClient()
 	cfg := util2.NewTestConfig()
@@ -391,9 +165,9 @@ func TestSocketForwarding(t *testing.T) {
 	os.Remove(cfg.Socket)
 	cfg.Status.ReportStatus = false
 
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	cli.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
 	go func() {
 		err := server.Run()
@@ -422,11 +196,11 @@ func TestSocket(t *testing.T) {
 	cfg.Host = ""
 	cfg.Status.ReportStatus = false
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	go func() {
 		err := server.Run()
 		require.NoError(t, err)
@@ -457,11 +231,11 @@ func TestSocketAndIp(t *testing.T) {
 	cfg.Port = cli.Port
 	cfg.Status.ReportStatus = false
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	cli.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
 	go func() {
 		err := server.Run()
@@ -622,11 +396,11 @@ func TestOnlySocket(t *testing.T) {
 	cfg.Host = "" // No network interface listening for mysql traffic
 	cfg.Status.ReportStatus = false
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	go func() {
 		err := server.Run()
 		require.NoError(t, err)
@@ -766,117 +540,14 @@ func TestOnlySocket(t *testing.T) {
 		})
 }
 
-// generateCert generates a private key and a certificate in PEM format based on parameters.
-// If parentCert and parentCertKey is specified, the new certificate will be signed by the parentCert.
-// Otherwise, the new certificate will be self-signed and is a CA.
-func generateCert(sn int, commonName string, parentCert *x509.Certificate, parentCertKey *rsa.PrivateKey, outKeyFile string, outCertFile string, opts ...func(c *x509.Certificate)) (*x509.Certificate, *rsa.PrivateKey, error) {
-	privateKey, err := rsa.GenerateKey(rand.Reader, 528)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-	notBefore := time.Now().Add(-10 * time.Minute).UTC()
-	notAfter := notBefore.Add(1 * time.Hour).UTC()
-
-	template := x509.Certificate{
-		SerialNumber:          big.NewInt(int64(sn)),
-		Subject:               pkix.Name{CommonName: commonName, Names: []pkix.AttributeTypeAndValue{util.MockPkixAttribute(util.CommonName, commonName)}},
-		DNSNames:              []string{commonName},
-		NotBefore:             notBefore,
-		NotAfter:              notAfter,
-		KeyUsage:              x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth, x509.ExtKeyUsageClientAuth},
-		BasicConstraintsValid: true,
-	}
-	for _, opt := range opts {
-		opt(&template)
-	}
-
-	var parent *x509.Certificate
-	var priv *rsa.PrivateKey
-
-	if parentCert == nil || parentCertKey == nil {
-		template.IsCA = true
-		template.KeyUsage |= x509.KeyUsageCertSign
-		parent = &template
-		priv = privateKey
-	} else {
-		parent = parentCert
-		priv = parentCertKey
-	}
-
-	derBytes, err := x509.CreateCertificate(rand.Reader, &template, parent, &privateKey.PublicKey, priv)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-
-	cert, err := x509.ParseCertificate(derBytes)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-
-	certOut, err := os.Create(outCertFile)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-	err = pem.Encode(certOut, &pem.Block{Type: "CERTIFICATE", Bytes: derBytes})
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-	err = certOut.Close()
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-
-	keyOut, err := os.OpenFile(outKeyFile, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-	err = pem.Encode(keyOut, &pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(privateKey)})
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-	err = keyOut.Close()
-	if err != nil {
-		return nil, nil, errors.Trace(err)
-	}
-
-	return cert, privateKey, nil
-}
-
-// registerTLSConfig registers a mysql client TLS config.
-// See https://godoc.org/github.com/go-sql-driver/mysql#RegisterTLSConfig for details.
-func registerTLSConfig(configName string, caCertPath string, clientCertPath string, clientKeyPath string, serverName string, verifyServer bool) error {
-	rootCertPool := x509.NewCertPool()
-	data, err := os.ReadFile(caCertPath)
-	if err != nil {
-		return err
-	}
-	if ok := rootCertPool.AppendCertsFromPEM(data); !ok {
-		return errors.New("Failed to append PEM")
-	}
-	clientCert := make([]tls.Certificate, 0, 1)
-	certs, err := tls.LoadX509KeyPair(clientCertPath, clientKeyPath)
-	if err != nil {
-		return err
-	}
-	clientCert = append(clientCert, certs)
-	tlsConfig := &tls.Config{
-		RootCAs:            rootCertPool,
-		Certificates:       clientCert,
-		ServerName:         serverName,
-		InsecureSkipVerify: !verifyServer,
-	}
-	return mysql.RegisterTLSConfig(configName, tlsConfig)
-}
-
 func TestSystemTimeZone(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	tk := testkit.NewTestKit(t, ts.store)
+	tk := testkit.NewTestKit(t, ts.Store)
 	cfg := util2.NewTestConfig()
 	cfg.Port, cfg.Status.StatusPort = 0, 0
 	cfg.Status.ReportStatus = false
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
 	defer server.Close()
 
@@ -885,9 +556,9 @@ func TestSystemTimeZone(t *testing.T) {
 }
 
 func TestInternalSessionTxnStartTS(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	se, err := session.CreateSession4Test(ts.store)
+	se, err := session.CreateSession4Test(ts.Store)
 	require.NoError(t, err)
 
 	_, err = se.Execute(context.Background(), "set global tidb_enable_metadata_lock=0")
@@ -915,16 +586,16 @@ func TestInternalSessionTxnStartTS(t *testing.T) {
 }
 
 func TestClientWithCollation(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	ts.RunTestClientWithCollation(t)
 }
 
 func TestCreateTableFlen(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	// issue #4540
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 	_, err = Execute(context.Background(), qctx, "use test;")
 	require.NoError(t, err)
@@ -994,9 +665,9 @@ func Execute(ctx context.Context, qc *server2.TiDBContext, sql string) (resultse
 }
 
 func TestShowTablesFlen(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 	ctx := context.Background()
 	_, err = Execute(ctx, qctx, "use test;")
@@ -1024,9 +695,9 @@ func checkColNames(t *testing.T, columns []*column.Info, names ...string) {
 }
 
 func TestFieldList(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 	_, err = Execute(context.Background(), qctx, "use test;")
 	require.NoError(t, err)
@@ -1107,24 +778,24 @@ func TestFieldList(t *testing.T) {
 }
 
 func TestClientErrors(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	ts.RunTestInfoschemaClientErrors(t)
 }
 
 func TestInitConnect(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	ts.RunTestInitConnect(t)
 }
 
 func TestSumAvg(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 	ts.RunTestSumAvg(t)
 }
 
 func TestNullFlag(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 
 	ctx := context.Background()
@@ -1195,10 +866,10 @@ func TestNullFlag(t *testing.T) {
 }
 
 func TestNO_DEFAULT_VALUEFlag(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	// issue #21465
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 
 	ctx := context.Background()
@@ -1218,7 +889,7 @@ func TestNO_DEFAULT_VALUEFlag(t *testing.T) {
 }
 
 func TestGracefulShutdown(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	cli := testserverclient.NewTestServerClient()
 	cfg := util2.NewTestConfig()
@@ -1227,7 +898,7 @@ func TestGracefulShutdown(t *testing.T) {
 	cfg.Status.StatusPort = 0
 	cfg.Status.ReportStatus = true
 	cfg.Performance.TCPKeepAlive = true
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
 	require.NotNil(t, server)
 	cli.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
@@ -1258,9 +929,9 @@ func TestGracefulShutdown(t *testing.T) {
 }
 
 func TestPessimisticInsertSelectForUpdate(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 	defer qctx.Close()
 	ctx := context.Background()
@@ -1282,7 +953,7 @@ func TestPessimisticInsertSelectForUpdate(t *testing.T) {
 }
 
 func TestTopSQLCatchRunningSQL(t *testing.T) {
-	ts := createTidbTestTopSQLSuite(t)
+	ts := servertestkit.CreateTidbTestTopSQLSuite(t)
 
 	db, err := sql.Open("mysql", ts.GetDSN())
 	require.NoError(t, err)
@@ -1346,13 +1017,13 @@ func TestTopSQLCatchRunningSQL(t *testing.T) {
 		stats := mc.GetSQLStatsBySQLWithRetry(query, true)
 		require.Greaterf(t, len(stats), 0, query)
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 	cancel()
 	wg.Wait()
 }
 
 func TestTopSQLCPUProfile(t *testing.T) {
-	ts := createTidbTestTopSQLSuite(t)
+	ts := servertestkit.CreateTidbTestTopSQLSuite(t)
 
 	db, err := sql.Open("mysql", ts.GetDSN())
 	require.NoError(t, err)
@@ -1441,7 +1112,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 			checkFn(ca.sql, ca.planRegexp)
 		}
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case 2: prepare/execute sql
 	cases2 := []struct {
@@ -1483,7 +1154,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 			checkFn(ca.prepare, ca.planRegexp)
 		}
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case 3: prepare, execute stmt using @val...
 	cases3 := []struct {
@@ -1532,7 +1203,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 			checkFn(ca.prepare, ca.planRegexp)
 		}
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case for other statements
 	cases4 := []struct {
@@ -1565,7 +1236,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 		// check for internal SQL.
 		checkFn("replace into mysql.global_variables (variable_name,variable_value) values ('tidb_stmt_summary_history_size', '5')", "")
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case for multi-statement.
 	cases5 := []string{
@@ -1584,7 +1255,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 			checkFn(sqlStr, ".*TableReader.*")
 		}
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case for multi-statement, but first statements execute failed
 	cases6 := []string{
@@ -1606,7 +1277,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 			require.Equal(t, 0, len(stats), sqlStr)
 		}
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case for multi-statement, the first statements execute success but the second statement execute failed.
 	cases7 := []string{
@@ -1624,7 +1295,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 	check = func() {
 		checkFn(cases7[0], "") // the first statement execute success, should have topsql data.
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case for statement with wrong syntax.
 	wrongSyntaxSQL := "select * froms t"
@@ -1637,7 +1308,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 		stats := mc.GetSQLStatsBySQL(wrongSyntaxSQL, false)
 		require.Equal(t, 0, len(stats), wrongSyntaxSQL)
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 
 	// Test case for high cost of plan optimize.
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/planner/mockHighLoadForOptimize", "return"))
@@ -1658,7 +1329,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 		require.Less(t, updateCPUTime, selectCPUTime)
 		checkFn(selectInPlanSQL, "")
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/planner/mockHighLoadForOptimize"))
 
 	// Test case for DDL execute failed but should still have CPU data.
@@ -1675,7 +1346,7 @@ func TestTopSQLCPUProfile(t *testing.T) {
 	check = func() {
 		checkFn(addIndexStr, "")
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/mockHighLoadForAddIndex"))
 
 	// Test case for execute failed cause by storage error.
@@ -1689,23 +1360,8 @@ func TestTopSQLCPUProfile(t *testing.T) {
 	check = func() {
 		checkFn(execFailedQuery, "")
 	}
-	ts.testCase(t, mc, execFn, check)
+	ts.TestCase(t, mc, execFn, check)
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/store/copr/handleTaskOnceError"))
-}
-
-func (ts *tidbTestTopSQLSuite) testCase(t *testing.T, mc *mockTopSQLTraceCPU.TopSQLCollector, execFn func(db *sql.DB), checkFn func()) {
-	var wg sync.WaitGroup
-	ctx, cancel := context.WithCancel(context.Background())
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		ts.loopExec(ctx, t, execFn)
-	}()
-
-	checkFn()
-	cancel()
-	wg.Wait()
-	mc.Reset()
 }
 
 func mustQuery(t *testing.T, dbt *testkit.DBTestKit, query string) {
@@ -2018,7 +1674,7 @@ func (c *resourceTagChecker) checkReqExist(t *testing.T, digest stmtstats.Binary
 	}
 }
 
-func setupForTestTopSQLStatementStats(t *testing.T) (*tidbTestSuite, stmtstats.StatementStatsMap, *resourceTagChecker, chan struct{}) {
+func setupForTestTopSQLStatementStats(t *testing.T) (*servertestkit.TidbTestSuite, stmtstats.StatementStatsMap, *resourceTagChecker, chan struct{}) {
 	// Prepare stmt stats.
 	stmtstats.SetupAggregator()
 
@@ -2037,7 +1693,7 @@ func setupForTestTopSQLStatementStats(t *testing.T) (*tidbTestSuite, stmtstats.S
 	})
 	stmtstats.RegisterCollector(mockCollector)
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	db, err := sql.Open("mysql", ts.GetDSN())
 	require.NoError(t, err)
@@ -2474,25 +2130,6 @@ func TestTopSQLResourceTag(t *testing.T) {
 	}
 }
 
-func (ts *tidbTestTopSQLSuite) loopExec(ctx context.Context, t *testing.T, fn func(db *sql.DB)) {
-	db, err := sql.Open("mysql", ts.GetDSN())
-	require.NoError(t, err, "Error connecting")
-	defer func() {
-		err := db.Close()
-		require.NoError(t, err)
-	}()
-	dbt := testkit.NewDBTestKit(t, db)
-	dbt.MustExec("use topsql;")
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-		}
-		fn(db)
-	}
-}
-
 func TestLocalhostClientMapping(t *testing.T) {
 	tempDir := t.TempDir()
 	socketFile := tempDir + "/tidbtest.sock" // Unix Socket does not work on Windows, so '/' should be OK
@@ -2503,11 +2140,11 @@ func TestLocalhostClientMapping(t *testing.T) {
 	cfg.Port = cli.Port
 	cfg.Status.ReportStatus = false
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	cli.Port = testutil.GetPortFromTCPAddr(server.ListenAddr())
 	go func() {
 		err := server.Run()
@@ -2596,7 +2233,7 @@ func TestLocalhostClientMapping(t *testing.T) {
 }
 
 func TestRcReadCheckTS(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	db, err := sql.Open("mysql", ts.GetDSN())
 	require.NoError(t, err)
@@ -2711,8 +2348,8 @@ func TestExtensionConnEvent(t *testing.T) {
 	})))
 	require.NoError(t, extension.Setup())
 
-	ts := createTidbTestSuite(t)
-	// createTidbTestSuite create an inner connection, so wait the previous connection closed
+	ts := servertestkit.CreateTidbTestSuite(t)
+	// servertestkit.CreateTidbTestSuite create an inner connection, so wait the previous connection closed
 	require.NoError(t, logs.waitEvent(extension.ConnDisconnected))
 
 	// test for login success
@@ -2827,8 +2464,8 @@ func TestExtensionConnEvent(t *testing.T) {
 }
 
 func TestSandBoxMode(t *testing.T) {
-	ts := createTidbTestSuite(t)
-	qctx, err := ts.tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
+	ts := servertestkit.CreateTidbTestSuite(t)
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
 	_, err = Execute(context.Background(), qctx, "create user testuser;")
 	require.NoError(t, err)
@@ -2861,7 +2498,7 @@ func TestSandBoxMode(t *testing.T) {
 // See: https://github.com/pingcap/tidb/issues/40979
 // Reusing memory of `chunk.Chunk` may cause some systems variable's memory value to be modified unexpectedly.
 func TestChunkReuseCorruptSysVarString(t *testing.T) {
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	db, err := sql.Open("mysql", ts.GetDSN())
 	require.NoError(t, err)
@@ -2898,7 +2535,7 @@ func TestChunkReuseCorruptSysVarString(t *testing.T) {
 }
 
 func TestTiDBIdleTransactionTimeout(t *testing.T) {
-	ts := createTidbTestTopSQLSuite(t)
+	ts := servertestkit.CreateTidbTestTopSQLSuite(t)
 	cases := []func(dbt *testkit.DBTestKit){}
 	// Test simple txn.
 	cases = append(cases, func(dbt *testkit.DBTestKit) {
@@ -3244,12 +2881,12 @@ func TestProxyProtocolWithIpFallbackable(t *testing.T) {
 	cfg.ProxyProtocol.Networks = "*"
 	cfg.ProxyProtocol.Fallbackable = true
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	// Prepare Server
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	go func() {
 		err := server.Run()
 		require.NoError(t, err)
@@ -3309,12 +2946,12 @@ func TestProxyProtocolWithIpNoFallbackable(t *testing.T) {
 	cfg.ProxyProtocol.Networks = "*"
 	cfg.ProxyProtocol.Fallbackable = false
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	// Prepare Server
-	server, err := server2.NewServer(cfg, ts.tidbdrv)
+	server, err := server2.NewServer(cfg, ts.Tidbdrv)
 	require.NoError(t, err)
-	server.SetDomain(ts.domain)
+	server.SetDomain(ts.Domain)
 	go func() {
 		err := server.Run()
 		require.NoError(t, err)
@@ -3348,10 +2985,10 @@ func TestConnectionWillNotLeak(t *testing.T) {
 	cfg.ProxyProtocol.Networks = "*"
 	cfg.ProxyProtocol.Fallbackable = false
 
-	ts := createTidbTestSuite(t)
+	ts := servertestkit.CreateTidbTestSuite(t)
 
 	cli := testserverclient.NewTestServerClient()
-	cli.Port = testutil.GetPortFromTCPAddr(ts.server.ListenAddr())
+	cli.Port = testutil.GetPortFromTCPAddr(ts.Server.ListenAddr())
 	dsn := cli.GetDSN(func(config *mysql.Config) {
 		config.User = "root"
 		config.DBName = "test"
@@ -3402,130 +3039,25 @@ func TestConnectionWillNotLeak(t *testing.T) {
 	}, time.Minute, time.Millisecond*100)
 }
 
-func TestCursorFetchErrorInFetch(t *testing.T) {
-	tmpStoragePath := t.TempDir()
-	restore := config.RestoreFunc()
-	defer restore()
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.TempStoragePath = tmpStoragePath
-	})
+func TestPrepareCount(t *testing.T) {
+	ts := servertestkit.CreateTidbTestSuite(t)
 
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	srv := server2.CreateMockServer(t, store)
-	srv.SetDomain(dom)
-	defer srv.Close()
-
-	appendUint32 := binary.LittleEndian.AppendUint32
-	ctx := context.Background()
-	c := server2.CreateMockConn(t, srv)
-
-	tk := testkit.NewTestKitWithSession(t, store, c.Context().Session)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(id int, payload BLOB)")
-	payload := make([]byte, 512)
-	for i := 0; i < 2048; i++ {
-		rand.Read(payload)
-		tk.MustExec("insert into t values (?, ?)", i, payload)
-	}
-
-	tk.MustExec("set global tidb_enable_tmp_storage_on_oom = ON")
-	tk.MustExec("set global tidb_mem_oom_action = 'CANCEL'")
-	defer tk.MustExec("set global tidb_mem_oom_action= DEFAULT")
-	require.Len(t, tk.Session().GetSessionVars().MemTracker.GetChildrenForTest(), 1)
-
-	// execute a normal statement, it'll spill to disk
-	stmt, _, _, err := c.Context().Prepare("select * from t")
+	qctx, err := ts.Tidbdrv.OpenCtx(uint64(0), 0, uint8(tmysql.DefaultCollationID), "test", nil, nil)
 	require.NoError(t, err)
-
-	tk.MustExec(fmt.Sprintf("set tidb_mem_quota_query=%d", 1))
-
-	require.NoError(t, c.Dispatch(ctx, append(
-		appendUint32([]byte{tmysql.ComStmtExecute}, uint32(stmt.ID())),
-		tmysql.CursorTypeReadOnly, 0x1, 0x0, 0x0, 0x0,
-	)))
-
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/util/chunk/get-chunk-error", "return(true)"))
-	defer func() {
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/util/chunk/get-chunk-error"))
-	}()
-	require.ErrorContains(t, c.Dispatch(ctx, appendUint32(appendUint32([]byte{tmysql.ComStmtFetch}, uint32(stmt.ID())), 1024)), "fail to get chunk for test")
-	// after getting a failed FETCH, the cursor should have been reseted
-	require.False(t, stmt.GetCursorActive())
-	require.Len(t, tk.Session().GetSessionVars().MemTracker.GetChildrenForTest(), 0)
-	require.Len(t, tk.Session().GetSessionVars().DiskTracker.GetChildrenForTest(), 0)
-}
-
-func TestCursorFetchShouldSpill(t *testing.T) {
-	restore := config.RestoreFunc()
-	defer restore()
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.TempStoragePath = t.TempDir()
-	})
-
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	srv := server2.CreateMockServer(t, store)
-	srv.SetDomain(dom)
-	defer srv.Close()
-
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/server/testCursorFetchSpill", "return(true)"))
-	defer func() {
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/server/testCursorFetchSpill"))
-	}()
-
-	appendUint32 := binary.LittleEndian.AppendUint32
+	prepareCnt := atomic.LoadInt64(&variable.PreparedStmtCount)
 	ctx := context.Background()
-	c := server2.CreateMockConn(t, srv)
-
-	tk := testkit.NewTestKitWithSession(t, store, c.Context().Session)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(id_1 int, id_2 int)")
-	tk.MustExec("insert into t values (1, 1), (1, 2)")
-	tk.MustExec("set global tidb_enable_tmp_storage_on_oom = ON")
-	tk.MustExec("set global tidb_mem_oom_action = 'CANCEL'")
-	defer tk.MustExec("set global tidb_mem_oom_action= DEFAULT")
-	require.Len(t, tk.Session().GetSessionVars().MemTracker.GetChildrenForTest(), 1)
-
-	// execute a normal statement, it'll spill to disk
-	stmt, _, _, err := c.Context().Prepare("select * from t")
+	_, err = Execute(ctx, qctx, "use test;")
 	require.NoError(t, err)
-
-	tk.MustExec(fmt.Sprintf("set tidb_mem_quota_query=%d", 1))
-
-	require.NoError(t, c.Dispatch(ctx, append(
-		appendUint32([]byte{tmysql.ComStmtExecute}, uint32(stmt.ID())),
-		tmysql.CursorTypeReadOnly, 0x1, 0x0, 0x0, 0x0,
-	)))
-}
-
-func TestCursorFetchExecuteCheck(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	srv := server2.CreateMockServer(t, store)
-	srv.SetDomain(dom)
-	defer srv.Close()
-
-	appendUint32 := binary.LittleEndian.AppendUint32
-	ctx := context.Background()
-	c := server2.CreateMockConn(t, srv)
-
-	stmt, _, _, err := c.Context().Prepare("select 1")
+	_, err = Execute(ctx, qctx, "drop table if exists t1")
 	require.NoError(t, err)
-
-	// execute with wrong ID
-	require.Error(t, c.Dispatch(ctx, append(
-		appendUint32([]byte{tmysql.ComStmtExecute}, uint32(stmt.ID()+1)),
-		tmysql.CursorTypeReadOnly, 0x1, 0x0, 0x0, 0x0,
-	)))
-
-	// execute with wrong flag
-	require.Error(t, c.Dispatch(ctx, append(
-		appendUint32([]byte{tmysql.ComStmtExecute}, uint32(stmt.ID())),
-		tmysql.CursorTypeReadOnly|tmysql.CursorTypeForUpdate, 0x1, 0x0, 0x0, 0x0,
-	)))
-
-	require.Error(t, c.Dispatch(ctx, append(
-		appendUint32([]byte{tmysql.ComStmtExecute}, uint32(stmt.ID())),
-		tmysql.CursorTypeReadOnly|tmysql.CursorTypeScrollable, 0x1, 0x0, 0x0, 0x0,
-	)))
+	_, err = Execute(ctx, qctx, "create table t1 (id int)")
+	require.NoError(t, err)
+	stmt, _, _, err := qctx.Prepare("insert into t1 values (?)")
+	require.NoError(t, err)
+	require.Equal(t, prepareCnt+1, atomic.LoadInt64(&variable.PreparedStmtCount))
+	require.NoError(t, err)
+	err = qctx.GetStatement(stmt.ID()).Close()
+	require.NoError(t, err)
+	require.Equal(t, prepareCnt, atomic.LoadInt64(&variable.PreparedStmtCount))
+	require.NoError(t, qctx.Close())
 }
